+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `reviewsHeading` and `reviewCommentsContainer` CSS handles.
+
 ## [1.5.1] - 2020-05-20
 
 ## [1.5.0] - 2020-04-29